Logic for the token is wrong
The use of token or basic auth was a conditional inside the TLS verify loop.
You can use TLS or skip the verification and use token or basic auth.

Just a tab issue. Put the token conditional outside the kubeInsecure check.

Also handled the case that we actually switch the boolean for skip TLS and forced it to replace true or false.