fix(rootfs): improve Makefiles for docker builds
This introduces a docker.mk file that now has shared logic between the
different builds. This will also create docker images that are stored in
non-public registries, which is great for cases where internet is
spotty or down. Some of the changes are just to normalize the Makefiles,
using more make and less arbitrary shell.

This partially addresses #292.
Closes #293.