fix(cli): revert to older deploy formatting
Prior to merging the projects, the version of `helm deploy` used a
different REST API (one without YAML in JSON). Post-merge, we need to
revert back to the older common.Template version.

As part of this, I switched an io.ReadCloser back to an io.Reader
because there is nowhere where the Closer part is used. The http lib
does not close a reader passed to it.