feat(helm): install directly from repo
A helm install command will try to load a chart from a local file first.
But if a local file is not found, it will try to fetch a file from a
matching repo request. The file will be downloaded to the client,
and then sent to Tiller for installation.