fix(tiller): allow 3rd party APIs
This feature has been disabled in the past because simply enabling the
feature causes the Kubernetes library to make requests to a server.
Thus, running any tests that use the 'pkg/kube' library has required
running a kube API server.

This patch makes it possible to selectively activate 3rdParty API
support, and then disables that support during testing.

Future versions of the Kubernetes library appear to make it easier to
configure and mock this behavior, so this is most likely a stop-gap
measure. (Famous last words.)

Closes #1472