Increase column width when listing releases
After #1560 was merged, names of releases can now be up to 53 characters long, up from 14 previously. When listing releases, the maximum column width is 30 characters, meaning it's not possible to get the full name of releases using the full available length.

This change just increases the maximum column width to 60, to allow for longer release names.