Link to the Helm Chart repository from the Helm project
The current README makes no mention of the Chart repository or where to find it. You have to Google to find it or stumble upon it. It was frustrating to come to the Helm project and not easily be able to find where the chart repository was - even though I knew it existed. This patch will make that easier for others.