feat(tiller): Stage 1 of new logging method
This is the first part of what will be several PRs to better the logging on Tiller.
This adds a `--log-level` flag with 3 configurable levels for Tiller and adds
logging for the `main` and `kube` packages.

To avoid creating one giant PR and to allow for feedback on the logging methodology
before doing it to everything, this only introduces the changes to the
aforementioned packages. 1 or 2 follow up PRs will add the new logging method
to all Tiller components and add documentation.