Restore the 'examples' directory
I'd included the /examples section as a subset of docs, and renamed the
section to chart-examples. However this breaks the CI tests, and in
hindsight they are best treated as development references rather than
a set of webpages, so I'll remove them from the site sidebar and restore
the directory name to fix the tests.