feat(*): add annotations to release
Annotations is an unstructured key value map stored with a release that
may be set by external tools to store and retrieve arbitrary metadata.

For example: currently there is no way to tell where the chart, from which
a release had been installed, came from (local file, repo, etc.). We can
store the chart's origin in the release annotations.

In future, when Helm will utilize TPRs or an addon API server to store
the release data, these annotations can easily be converted to the k8s
API's v1.ObjectMeta.Annotations.