feat(functional_tests): add functional tests
This adds a series of functional tests that are implemented as Bash
scripts and run inside of a Docker container.

The idea is to create an intuitive but useful set of functional tests
that we can run against an existing Kubermetes cluster as acceptance
tests for a Helm release.

Scripts are spaced out and organized thematically, with basic tests
preceding more sophisticated tests. Emphasis falls on testing different
Kubernetes resource kinds, as well as testing all of Helm's commands.

The test runs with configmap storage since that is the default, even
though in-memory storage would be faster and easier to clean up.

The test suite is run inside of a namespace, and the namespace is
destroyed at the end of a test run.