Add optional check to restrict charts to target namespace
This change adds a new `RestrictToNamespace` option to gRPC calls.
When enabled, this will reject charts that attempt to install
resources outside the target namespace (either in another namespace,
or "global" non-namespaced resources).

This is intended as a simple namespace-scoped stop-gap permission
model, until helm/tiller becomes more RBAC-aware.  It implements a
simple binary mechanism that allows either "namespace local" or
"admin-installed global" operations, and trusts the client to indicate
which operation was intended.  It would be reasonable to think of this
as "the sudo approach".

Note:
- This change does not (yet) expose this option through to the `helm`
  CLI tool.
- This change does not (yet) modify rudder.  Rudder continues to allow
  any namespace in chart resources.