Ask ioutil to write in a temp file
Similar to https://github.com/kubernetes/helm/issues/3253 we're using
Helm by directly importing the Go package.

A concurrency issue arises when multiple requests come for the same
package. Currently DownloadTo truncates contents of destination file and
rewrites it.

This patch changes the behaviour so instead of directly writing into the
destination it'll write the content in a temp file and renames it once
download is finished.

This change fixes sporadic issues in
https://github.com/amir/skeg/blob/5c5c6a1c215cf5110843a28f1fa389aa69a3f452/main_test.go