fix: bring back Release.Namespace
any chart which uses `RoleBinding` or `ClusterRoleBinding` needs a way to populate the namespace being used for a release - such as to reference a `ServiceAccount` in a `RoleBinding`. Removing support for `Release.Namespace` breaks many charts.

Given the namespace being used to install/update a chart is typically fixed; am not sure there's a great reason for its removal - as its totally repeatable.

fixes #4255