fix(helm): Update status output to include resource details
Update of the client-go package changed the status output to only include
the age of resources. The new printer in client-go only formats the output
to include details of specific resources if the internal representation
of resources are passed into the printer. This PR updates helm to convert
resources to the internal type before printing.

Closes #4712

Signed-off-by: Morten Torkildsen <mortent@google.com>