fix(helm): Use line breaks consistently in status output
The output from helm status does not have consistent use of line breaks.
For some resources there is a line break after the kind header, for
others there is not. This is caused by how the printer handles column
headers. This removes a line break for all but the first resource listed.

Signed-off-by: Morten Torkildsen <mortent@google.com>