fix(helm): Print details for pod resource
Due to a regression from a previous change, details about pod resources
does not show up in the status output. This makes sure that the pod type
from core are passed in to the printer so the details are shown in the
output.

Signed-off-by: Morten Torkildsen <mortent@google.com>
(cherry picked from commit ede43a313dde2611ea63e89546af46854b4121d4)