fix(helm): Allow custom resources in hooks (#4986)
Currently the code that handles hooks uses a builder that creates the versioned types rather than unstructured. This results in an error whenever a custom resource is used in the hook as the type will not be registered in the scheme used in Helm. This changes this to use a builder that created unstructured resources and only converts to the versioned type when needed.

Signed-off-by: Morten Torkildsen <mortent@google.com>
Signed-off-by: Kevin Labesse <kevin@labesse.me>