feat(tiller): add disk storage method
There are issues with the configmap and secrets storage when releases
are becoming larger than 1mb, since Kubernetes (etcd) doesn't support
entries larger than that.
This tries to solve that problem by adding a new storage type disk.
This will store the release to the local disk, in order to keep it
persistent.

Signed-off-by: Soren Mathiasen <smo@tradeshift.com>