Try a bit harder to provide a clear error message.
When an unexpected file format is provided for a chart archive, spend
some extra effort to provide the user with some meaningful feedback.

This is intended to help users who mistakenly invoke `helm template`
like this:

    helm template values.yaml charts/mychart

This is a user error, and now we can report that the argument
`values.yaml` was expected to be a chart.

Fixes #5296

Signed-off-by: Jordan Sissel <jls@semicomplete.com>