Fix debug printouts for zsh completion
Cobra provides some out-of-the-box debugging for bash completion.
To use it, one must set the variable BASH_COMP_DEBUG_FILE to
some file where the debug output will be written.  Many of the
debug printouts indicate the current method name; they do so
by using bash's ${FUNCNAME[0]} variable. This variable is
different in zsh. To obtain the current method name in zsh
we must use ${funcstack[1]}.

This commit adds the proper sed modification to convert from
bash to zsh.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>