Cannot have an = sign as an index
Flags sometimes can be used with an = sign, such as --kube-context=prod.
In this case, the variable ${flagname} retains the = sign as part of the
flag name. However, in zsh completion, an = sign cannot be part of an
index of the associative array 'flaghash' or else it causes an error.

This commits strips the = sign out when using ${flagname} as an index.

Note that this is not a big deal since flaghash is not actually used
anywhere in Helm completion. I believe it is made available by the
Cobra framework in case some completions choose to use it.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>