fix(helm): update Cobra to latest version
The version of Cobra used in dev-v3 is older than the one used
for the master branch (helm v2).  Although the dev-v3 branch
was based on master and therefore had the same Cobra version
originally, it was changed a couple of times to choose Cobra
tagged versions instead.

However, the latest Cobra tagged version is older than the version
of Cobra used on the master branch.  Therefore, some of the
improvements made to Cobra and used by helm v2 are not available
to helm v3 currently.

This commit brings Cobra to its latest available commit for
helm v3.  Originally, I thought of selecting the same Cobra version
in v3 as for helm v2, but after reviewing the 5 extra commits
added to Cobra, I feel it is safe and valuable to bring them in
also.

Bringing Cobra up-to-date is essential for upcoming work being
prepared for dynamic bash-completion; there are bug fixes in
Cobra that are necessary for dynamic bash-completion to work
properly.  Specifically,
    https://github.com/spf13/cobra/pull/730
which fixes
    https://github.com/spf13/cobra/issues/694
is essential to avoid the risk of colliding and possibly
breaking kubectl dynamic bash-completion once helm v3 has its
own dynamic completion.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>