(helm): Proper fix for #5046
PR #5072 followed by #5406 tweaked the handling of the associative
array 'aliashash' when in zsh.  However, upon further investigation
the root of the problem was the 'aliashash' was not being declared
properly as it was declared within a method and therefore not
accessible to the rest of the completion script.

The previous commit of this PR makes the necessary change to properly
declare 'aliashash' which makes the previous tweak unecessary.  This
commit removes the tweak.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>