Allow for completion of aliases
Turns out that removing the quotes from the associate array
'aliashash' is still required, but because it is needed for
completion of alias commands.  For example
  helm dep <TAB>
does not complete as expected (as if it was 'helm dependency')
if the quotes are not removed.

This is because although bash ignores quotes when using
associative arrays, zsh does not.  So when looking for an alias
  aliashash[dep]
will not match the entry
  aliashash["dep"]
in zsh but will for bash.  Therefore, removing the quotes fixes
things.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>