ref(rollback): consolidate `helm rollback` logic to re-use `helm upgrade`
A rollback is an "upgrade" of an older revision. This change removes all
the one-off logic from `helm rollback`, relying on the business logic backing
`helm upgrade`.

This comes with several advantages:

- uniform behaviour with regards to feature flags
- features implemented for `helm upgrade` can be enabled for `helm rollback` with no significant development cost

This change removes the `pre-rollback` and `post-rollback` hooks, opting to use the existing `pre-upgrade` and `post-upgrade` flags. These flags are not implemented by any of the charts in the incubator or stable repositories, so it appears safe to remove.

Signed-off-by: Matthew Fisher <matt.fisher@microsoft.com>