fix: upgrade with CRD changes
Probably since K8s 1.13.x, `converter.ConvertToVersion(info.Object, groupVersioner)` which is the body of `asVersioned` doesn't return an error or an "unstructured" object, but `apiextensions/v1beta1.CustomResourceDefinition`.

The result was `helm upgrade` with any changes in CRD consistently failing.

This fixes that by adding an additional case of the conversion result being `v1beta1.CustomResourceDefinition`.

This is a backward-compatible change as it doesn't remove existing switch cases for older K8s versions.

Fixes #5853

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>