Add tests for helm completion (#6115)
This commit adds a framework and tests for helm command-line
completion.

To cover different shells (bash 4, bash 3, zsh) we use docker.

For MacOS, the tests are run locally if the host is MacOS and has
the right setup (bash completion installed, or zsh installed).

Dynamic-completion (e.g., helm status <TAB>) is not yet tested, but has
been considered in the design of the framework.

To run the tests:
  make test-completion

Although the framework supports marking tests as known failures,
this commit commented out the failing tests to allow for a clean
output for this proposal.  They can be uncommented later on.

A new directory is added: scripts/completion-tests.
It contains the required scripts for the completion tests.

scripts/completion-tests/completionTests.sh:
   Helm-specific completion tests.
   This script should evolve as new tests are added and new
   completion features implemented.

scripts/completion-tests/test-completion.sh:
   This script repeatedly runs the completion-tests in different
   environments using docker.
   This script would need changes if a new environment needs to
   be added.

scripts/completion-tests/lib/completionTests-base.sh:
   The magic. It is this script that allows completion testing.
   This script is not expected to change in day-to-day helm
   evolution.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>