Clarify operator docs wording
The original is explaining the negation, when the body would not be
included. Which would happen in the complement of the if expression,
ie. flipped by De Morgan's law's:

```
not (or .Values.anUnsetVariable (not .Values.aSetVariable))
==
and (not .Values.anUnsetVariable) .Values.aSetVariable
```

> unset variables evaluate to false

is equivalent to `not .Values.anUnsetVariable`.

> and

is equivalent to `and`.

> .Values.setVariable was negated with the not function

doesn't seem to match `.Values.aSetVariable`.
To me, that would be `not .Values.aSetVariable` instead.

Anyway, explaining the `if` expression as-is and not the negation is
clearer and parallels the first `if` operator.

Signed-off-by: Kevin Lau <kelau1993@gmail.com>