Dynamic completion to use same binary as main call
The binary of Helm to use for dynamic completion should be the same
as the actual Helm binary being used.  For example, if PATH points
to a version of helm v2, but the user calls bin/helm to use a local
v3 version, then dynamic completion should also use bin/helm.
If not, in this example, the dynamic completion will use the
information returned by helm v2.

This improvement is particularly useful for users that will run both
helm v2 and helm v3 at the same time.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>