Create file locking directory if it does not exist
As Helm v3 uses lazy creation for configuration then
directories and files are not created until required.
File locking when doing repo add was introduced in
v2 and ported to v3 in #6492. It locks on the config
directory where the repo file resides and therefore
needs the directory to be created if it doesn't exist.
This fix adds the directory if need be,

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>