feat(cmd): Replace 'helm show' with 'helm show all'
As part of #6552

The is a break in compatibility because 'helm show' is no longer a
valid command on its own; what it used to do is now achieved with
'helm show all'.

This change avoids confusion between chart reference and subcommands.
It also opens the door to dynamic shell comnpletion.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>