fix(pkg/downloader): Implement ensureLock() instead of digest comparison
The functionality of checking whether the lock file is out-of-date or not is
quite important to end-users. But digest comparison is not an answer for it.
This commit implements it in an alternative way, which compares names, repos,
and version resolutions in each pair of Chart.yaml and Chart.lock dependencies.

Signed-off-by: Hang Park <hangpark@kaist.ac.kr>