Fixing regression for numeric values refs #6708
This commit is a fix for the regression introduced in 2.15
with an attempt to carry all numeric values in json.Number container
(see https://github.com/helm/helm/issues/6708 for more details).

The fix overloads a series of Helm template functions:
  * int64
  * int
  * float64
  * add1
  * add
  * sub
  * div
  * mod
  * mul
  * max
  * min
  * ceil
  * floor
  * round
  * until
  * untilStep
  * splitn
  * abbrev
  * abbrevboth
  * trunc
  * substr
  * repeat
  * randAlphaNum
  * randAlpha
  * randAscii
  * randNumeric
  * wrap
  * wrapWith
  * indent
  * nindent
  * plural
  * slice
  * eq
  * gt
  * ge
  * lt
  * le
  * ne

Corresponding test cases are included in the engine test suite.

Signed-off-by: Oleg Sidorov <oleg.sidorov@booking.com>