fix(kube): Fixes wait for jobs
This happened to be a bug we identified in Helm 3 and did not check if
it existed in Helm 2. The improved logic for job waiting used an automatic
retry. However, when we were creating the watcher, we were listing on everything
of that same api version and kind. So if you had more than 1 hook and the first
was successful, it would think everything was successful. I have validated that
this now fails as intended if a job is failing

Closes #6767

Signed-off-by: Taylor Thomas <taylor.thomas@microsoft.com>