fix(get): install Helm v2.16.1
This is a temporary fix. This prevents the get script from installing Helm 3 as soon as it's released, potentially causing disruption to users that were expecting a Helm 2 release.

A `get-helm-3` script is also supplied, which is identical to the previous `get` script. That way, Helm 3 users also have an equivalent way to install Helm 3.

Signed-off-by: Matthew Fisher <matt.fisher@microsoft.com>