feat(comp): Dynamic completion for plugins
If the completion script determines there are no known completions
it will look for a plugin that matches the command line.  If found,
the completion script will call the plugin through helm with the
special __complete__ parameter.  Helm will detect that special
parameter and then call an executable called plugin.complete which is
optionally provided by the plugin in its root directory.

Helm will set the plugin environment variable and pass the entire
current command line to that script file.
That executable file should output the possible completions
based on the command-line it received.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>