feat(helm): add recreate upgrade (rollback) strategy
An additional optional flag --recreate can be passed on upgrade (or rollback) of a release.
In combination with the --force flag the following strategies are employed when updating a resource (which can be directly compared to kubectl):

helm                        kubectl          action                         on 'invalid' or 'conflict'
--------------------------------------------------------------------------------------------------------------

upgrade                     apply            PATCH                          fail

upgrade --force             replace          PUT                            fail

upgrade --recreate          apply --force    PATCH                          DELETE -> GET (poll) -> POST

upgrade --recreate --force  replace --force  DELETE -> GET (poll) ->  POST  fail

The 'on error' column should be interpreted as follows. The server responds with 'invalid' e.g. if a certain resource contains an immutable field, which cannot be patched or updated by a PUT. In theses cases it can be helpful to delete the resource and recreate it. Examples for theses cases are:

 * roleRef in RoleBinding
 * spec.clusterIP in Service
 * parameters in StorageClass

Closes #7082.

Signed-off-by: Daniel Strobusch <1847260+dastrobu@users.noreply.github.com>