Workaround obsolete API versions
This is a workaround for changes in API versions like
extensions/v1beta1 => apps/v1

Problem:
From Helm's perspective those are new Resources as the API version
cannot be changed. However there is a compat layer in k8s that will
return the old resources for the new API version. Helm will reject the
resource 'creation' as there is already a resource of the same kind and
version.

Solution:
In general ignore the Version.
Due to the fact that there might be clashes of Kind we track CRD and
none-CRD resources seperate.

Open issue:
Corner case with same Kind in different CRDs and a pre-existing resource
that would clash with the one created by Helm.

Signed-off-by: Sebastian Poehn <sebastian.poehn@gmail.com>