Fixing issue where archives created on windows have broken paths
When archives are created on windows the path spearator in the
archive file is \\. This causes issues when the file is unpacked.
For example, on Linux the files are unpacked in a flat structure
and \ is part of the file name. This causes comp issues. In Helm
v2 the path was set as / when the archive was written. This works
on both Windows and POSIX systems.

The fix being implemented is to use the ToSlash function to ensure
/ is used as the separator.

Fixes #7748

Signed-off-by: Matt Farina <matt@mattfarina.com>
(cherry picked from commit 28b085bed5d35d6495e06d4fc7763f4028142560)