Experimental Notary Support for Helm
Notary v1 is the initial implementation of theUpdateFramework.
This framework basically helps understand the security issues,
that come up when updating a repo and addresses them by
adding multilevel signature checking.

Using Notary OCI Artifacts can also be provenance checked. and
Helm Charts can also benefit from it. With that in mind,
there needs to be an interface between pushing and pulling the
charts, to and from the remote registries, to check if the right
chart is being downloaded.

Here --sign is used with helm chart pull/push to verify/set
signatures in Notary.

--ca-cert is given to set the CA Cert
--trust-server would be the Notary Server
--trust-dir would be the directory used instead of the trust server

Co-authored-by: Radu Matei <root@radu.sh>
Co-authored-by: Vibhav Bobade <vibhav.bobde@gmail.com>
Signed-off-by: Vibhav Bobade <vibhav.bobde@gmail.com>